+++ src/projects/components/ProjectDashboard.tsx
@@ -1,16 +1,12 @@
 import { useEffect } from "react"
-import { Routes, useParam } from "@blitzjs/next"
+import { useParam } from "@blitzjs/next"
 import { useMutation, useQuery } from "@blitzjs/rpc"
 import { TaskStatus } from "db"
 import moment from "moment"
-import Link from "next/link"
-import Table from "src/core/components/Table"
 import getTasks from "src/tasks/queries/getTasks"
 import { useCurrentUser } from "src/users/hooks/useCurrentUser"
 import getProjectStats from "../queries/getProjectStats"
 import getContributors from "src/contributors/queries/getContributors"
-import { UserIcon, GlobeAltIcon, ArchiveBoxIcon } from "@heroicons/react/24/outline"
-import { CircularProgressbar, buildStyles } from "react-circular-progressbar"
 import "react-circular-progressbar/dist/styles.css"
 import React, { useState } from "react"
 import {
@@ -30,55 +26,21 @@ import getProjectWidgets from "src/widgets/queries/getProjectWidgets"
 import toast from "react-hot-toast"
 import getProject from "src/projects/queries/getProject"
 import getNotifications from "src/notifications/queries/getNotifications"
-import {
-  notificationColumns,
-  projectManagersColumns,
-  projectTaskColumns,
-} from "src/widgets/components/ColumnHelpers"
-import DateFormat from "src/core/components/DateFormat"
+import useDashboardDragHandlers from "src/widgets/hooks/useDashboardDragHandlers"
 
 const ProjectDashboard = () => {
-  //default information
-  const projectId = useParam("projectId", "number")
+  const [updateWidgetMutation] = useMutation(updateProjectWidgets)
+
   const currentUser = useCurrentUser()
+
+  const { handleDragEnd } = useDashboardDragHandlers({
+    setBoxes,
+    updateWidgetMutation,
+  })
+
   const today = moment().startOf("minute")
+  const projectId = useParam("projectId", "number")
   const [project] = useQuery(getProject, { id: projectId })
-
-  // dragging information
-  const handleDragEnd = async (event) => {
-    const { active, over } = event
-    if (over && active.id !== over.id) {
-      setBoxes((currentBoxes) => {
-        const oldIndex = currentBoxes.findIndex((box) => box.id === active.id)
-        const newIndex = currentBoxes.findIndex((box) => box.id === over.id)
-        const newBoxes = arrayMove(currentBoxes, oldIndex, newIndex)
-
-        // Update positions based on new order in the state
-        const updatedPositions = newBoxes.map((box, index) => ({
-          id: box.id,
-          position: index + 1,
-        }))
-
-        // Call the mutation
-        updateWidgetMutation({ positions: updatedPositions })
-          .then(() => {
-            //console.log("Widget positions updated successfully")
-          })
-          .catch((error) => {
-            //console.error("Error updating widget positions:", error)
-          })
-
-        return newBoxes
-      })
-    }
-  }
-  const sensors = useSensors(
-    useSensor(PointerSensor),
-    useSensor(TouchSensor),
-    useSensor(KeyboardSensor, {
-      coordinateGetter: sortableKeyboardCoordinates,
-    })
-  )
 
   // Get the widgets for the user
   const [boxes, setBoxes] = useState([])
@@ -88,237 +50,8 @@ const ProjectDashboard = () => {
   })
 
   // mutations for the widgets
-  const [updateWidgetMutation] = useMutation(updateProjectWidgets)
+
   const [setWidgetMutation] = useMutation(setProjectWidgets)
-
-  // links
-  const projectLink = (
-    <Link
-      className="btn btn-primary self-end m-4"
-      href={Routes.EditProjectPage({ projectId: projectId! })}
-    >
-      Edit Project
-    </Link>
-  )
-  const taskLink = (
-    <Link
-      className="btn btn-primary self-end m-4"
-      href={Routes.TasksPage({
-        projectId: projectId,
-      })}
-    >
-      All Tasks
-    </Link>
-  )
-  const notificationLink = (
-    <Link
-      className="btn btn-primary self-end m-4"
-      href={Routes.ProjectNotificationsPage({ projectId: projectId! })}
-    >
-      Project Notifications
-    </Link>
-  )
-  const contributorLink = (
-    <Link
-      className="btn btn-primary self-end m-4"
-      href={Routes.ContributorsPage({ projectId: projectId! })}
-    >
-      View
-    </Link>
-  )
-  const teamLink = (
-    <Link
-      className="btn btn-primary self-end m-4"
-      href={Routes.TeamsPage({ projectId: projectId! })}
-    >
-      View
-    </Link>
-  )
-  const formLink = (
-    <Link
-      className="btn btn-primary self-end m-4"
-      href={Routes.MetadataPage({ projectId: projectId! })}
-    >
-      View
-    </Link>
-  )
-  const elementLink = (
-    <Link
-      className="btn btn-primary self-end m-4"
-      href={Routes.ElementsPage({ projectId: projectId! })}
-    >
-      View
-    </Link>
-  )
-  const labelLink = (
-    <Link
-      className="btn btn-primary self-end m-4"
-      href={Routes.CreditPage({ projectId: projectId! })}
-    >
-      View
-    </Link>
-  )
-  const taskSummaryLink = (
-    <Link
-      className="btn btn-primary self-end m-4"
-      href={Routes.TasksPage({
-        projectId: projectId,
-      })}
-    >
-      View
-    </Link>
-  )
-
-  // displays
-  const getProjectDisplay = (project) => {
-    return (
-      <div>
-        {project.description}
-        <p className="italic">
-          Last update: <DateFormat date={project.updatedAt}></DateFormat>
-        </p>
-
-        <p className="font-bold mt-4">Contacts for the Project: </p>
-        <Table
-          columns={projectManagersColumns}
-          data={projectManagers}
-          classNames={{
-            thead: "text-sm",
-            tbody: "text-sm",
-            td: "text-sm",
-          }}
-        />
-      </div>
-    )
-  }
-  const getUpcomingTaskDisplay = (upcomingTasks) => {
-    if (upcomingTasks.length === 0) {
-      return <p className="italic p-2">No upcoming tasks</p>
-    }
-
-    return (
-      <Table
-        columns={projectTaskColumns}
-        data={upcomingTasks.slice(0, 3)}
-        classNames={{
-          thead: "text-sm text-base-content",
-          tbody: "text-sm text-base-content",
-          td: "text-sm text-base-content",
-        }}
-      />
-    )
-  }
-  const getOverdueTaskDisplay = (pastDueTasks) => {
-    if (pastDueTasks.length === 0) {
-      return <p className="italic p-2">No overdue tasks</p>
-    }
-    return (
-      <Table
-        columns={projectTaskColumns}
-        data={pastDueTasks.slice(0, 3)}
-        classNames={{
-          thead: "text-sm text-base-content",
-          tbody: "text-sm text-base-content",
-          td: "text-sm text-base-content",
-        }}
-      />
-    )
-  }
-  const getNotificationDisplay = (notifications) => {
-    if (notifications.length === 0) {
-      return <p className="italic p-2">No unread notifications</p>
-    }
-
-    return (
-      <Table
-        columns={notificationColumns}
-        data={notifications}
-        classNames={{
-          thead: "text-sm text-base-content",
-          tbody: "text-sm text-base-content",
-          td: "text-sm text-base-content",
-        }}
-      />
-    )
-  }
-  const getContributorDisplay = (projectStats) => {
-    return (
-      <div className="flex justify-center font-bold text-3xl">
-        {projectStats.allContributor}
-        <UserIcon className="w-20" />
-      </div>
-    )
-  }
-  const getTeamDisplay = (projectStats) => {
-    return (
-      <div className="flex justify-center font-bold text-3xl">
-        {projectStats.allTeams}
-        <GlobeAltIcon className="w-20" />
-      </div>
-    )
-  }
-  const getFormDisplay = (projectStats) => {
-    return (
-      <div className="flex justify-center font-bold text-3xl">
-        <CircularProgressbar
-          value={formPercent * 100}
-          text={`${Math.round(formPercent * 100)}%`}
-          styles={buildStyles({
-            textSize: "16px",
-            pathTransitionDuration: "none",
-            pathColor: "oklch(var(--p))",
-            textColor: "oklch(var(--s))",
-            trailColor: "oklch(var(--pc))",
-            backgroundColor: "oklch(var(--b3))",
-          })}
-        />
-      </div>
-    )
-  }
-  const getTotalTaskDisplay = (projectStats) => {
-    return (
-      <div className="flex justify-center font-bold text-3xl">
-        <CircularProgressbar
-          value={taskPercent * 100}
-          text={`${Math.round(taskPercent * 100)}%`}
-          styles={buildStyles({
-            textSize: "16px",
-            pathTransitionDuration: "none",
-            pathColor: "oklch(var(--p))",
-            textColor: "oklch(var(--s))",
-            trailColor: "oklch(var(--pc))",
-            backgroundColor: "oklch(var(--b3))",
-          })}
-        />
-      </div>
-    )
-  }
-  const getElementDisplay = (projectStats) => {
-    return (
-      <div className="flex justify-center font-bold text-3xl">
-        {projectStats.allElements}
-        <ArchiveBoxIcon className="w-20" />
-      </div>
-    )
-  }
-  const getLabelsDisplay = (projectStats) => {
-    return (
-      <div className="flex justify-center font-bold text-3xl">
-        <CircularProgressbar
-          value={labelPercent * 100}
-          text={`${Math.round(labelPercent * 100)}%`}
-          styles={buildStyles({
-            textSize: "16px",
-            pathTransitionDuration: "none",
-            pathColor: "oklch(var(--p))",
-            textColor: "oklch(var(--s))",
-            trailColor: "oklch(var(--pc))",
-            backgroundColor: "oklch(var(--b3))",
-          })}
-        />
-      </div>
-    )
-  }
 
   //get the data
   // get the project manangers
@@ -416,136 +149,23 @@ const ProjectDashboard = () => {
       const sortedWidgets = fetchedWidgets.sort((a, b) => a.position - b.position)
       const updatedBoxes = sortedWidgets.map((widget) => {
         switch (widget.type) {
-          case "ProjectSummary":
-            return {
-              id: widget.id,
-              title: project.name,
-              display: getProjectDisplay(project),
-              link: projectLink,
-              position: widget.position,
-              size: "col-span-6",
-              tooltipId: "tool-project",
-              tooltipContent: "Overall project information",
-            }
-          case "Notifications":
-            return {
-              id: widget.id,
-              title: "Notifications",
-              display: getNotificationDisplay(notifications),
-              link: notificationLink,
-              position: widget.position,
-              size: "col-span-6",
-              tooltipId: "tool-notification",
-              tooltipContent: "Three notifications for this project",
-            }
-          case "OverdueTask":
-            return {
-              id: widget.id,
-              title: "Overdue Tasks",
-              display: getOverdueTaskDisplay(pastDueTasks),
-              link: taskLink,
-              position: widget.position,
-              size: "col-span-6",
-              tooltipId: "tool-overdue",
-              tooltipContent: "Three overdue tasks for this project",
-            }
-          case "UpcomingTask":
-            return {
-              id: widget.id,
-              title: "Upcoming Tasks",
-              display: getUpcomingTaskDisplay(upcomingTasks),
-              link: taskLink,
-              position: widget.position,
-              size: "col-span-6",
-              tooltipId: "tool-upcoming",
-              tooltipContent: "Three upcoming tasks for this project",
-            }
-          case "ContributorNumber":
-            return {
-              id: widget.id,
-              title: "Contributors",
-              display: getContributorDisplay(projectStats),
-              link: contributorLink,
-              position: widget.position,
-              size: "col-span-2",
-              tooltipId: "tool-contributors",
-              tooltipContent: "Total number of contributors",
-            }
-          case "TeamNumber":
-            return {
-              id: widget.id,
-              title: "Teams",
-              display: getTeamDisplay(projectStats),
-              link: teamLink,
-              position: widget.position,
-              size: "col-span-2",
-              tooltipId: "tool-teams",
-              tooltipContent: "Total number of teams",
-            }
-          case "FormNumber":
-            return {
-              id: widget.id,
-              title: "Forms",
-              display: getFormDisplay(projectStats),
-              link: formLink,
-              position: widget.position,
-              size: "col-span-2",
-              tooltipId: "tool-forms",
-              tooltipContent: "Percent of forms completed",
-            }
-          case "TaskTotal":
-            return {
-              id: widget.id,
-              title: "Tasks",
-              display: getTotalTaskDisplay(projectStats),
-              link: taskSummaryLink,
-              position: widget.position,
-              size: "col-span-2",
-              tooltipId: "tool-tasks",
-              tooltipContent: "Percent of tasks completed",
-            }
-          case "ElementSummary":
-            return {
-              id: widget.id,
-              title: "Elements",
-              display: getElementDisplay(projectStats),
-              link: elementLink,
-              position: widget.position,
-              size: "col-span-2",
-              tooltipId: "tool-element",
-              tooltipContent: "Number of elements for this project",
-            }
-          case "LabelsSummary":
-            return {
-              id: widget.id,
-              title: "Roles",
-              display: getLabelsDisplay(projectStats),
-              link: labelLink,
-              position: widget.position,
-              size: "col-span-2",
-              tooltipId: "tool-labels",
-              tooltipContent: "Percent of contributors or tasks labeled",
-            }
-          default:
-            return {
-              id: widget.id,
-              title: "Unknown Widget",
-              display: <div>Widget configuration error</div>,
-              link: <div />,
-              position: widget.position,
-              size: "col-span-4",
-              tooltipId: "tool-unknown",
-              tooltipContent: "Unknown widget",
-            }
         }
       })
       setBoxes(updatedBoxes)
     }
   }, [fetchedWidgets])
 
+  const sensors = useSensors(
+    useSensor(PointerSensor),
+    useSensor(TouchSensor),
+    useSensor(KeyboardSensor, {
+      coordinateGetter: sortableKeyboardCoordinates,
+    })
+  )
+
   return (
     <div className="flex flex-col space-y-4">
-      <DndContext collisionDectection={closestCorners} onDragEnd={handleDragEnd} sensors={sensors}>
+      <DndContext collisionDetection={closestCorners} onDragEnd={handleDragEnd} sensors={sensors}>
         <SortableBox boxes={boxes} />
       </DndContext>
     </div>
