+++ src/widgets/utils/widgetRegistry.tsx
@@ -1,0 +1,152 @@
+import PrimaryLink from "src/core/components/PrimaryLink"
+import {
+  GetProjectDisplay,
+  GetUpcomingTaskDisplay,
+  GetOverdueTaskDisplay,
+  GetNotificationDisplay,
+  GetProjectSummaryDisplay,
+  GetProjectOverdueTaskDisplay,
+  GetProjectUpcomingTaskDisplay,
+  GetContributorDisplay,
+  GetTeamDisplay,
+  GetFormDisplay,
+  GetTotalTaskDisplay,
+  GetElementDisplay,
+  GetLabelsDisplay,
+} from "src/core/components/GetDashboardDisplay"
+import { Routes } from "@blitzjs/next"
+
+export const widgetRegistry = {
+  LastProject: {
+    title: "Last Updated Projects",
+    display: GetProjectDisplay,
+    linkRoute: Routes.ProjectsPage(),
+    linkText: "All Projects",
+    size: "col-span-6",
+    tooltipId: "tool-last-project",
+    tooltipContent: "Three recently updated projects",
+  },
+  Notifications: {
+    title: "Notifications",
+    display: GetNotificationDisplay,
+    linkRoute: Routes.NotificationsPage(),
+    linkText: "All Tasks",
+    size: "col-span-6",
+    tooltipId: "tool-notifications",
+    tooltipContent: "Three recent notifications for all projects",
+  },
+  OverdueTask: {
+    title: "Overdue Tasks",
+    display: GetOverdueTaskDisplay,
+    linkRoute: Routes.AllTasksPage(),
+    linkText: "All Notifications",
+    size: "col-span-6",
+    tooltipId: "tool-overdue",
+    tooltipContent: "Three overdue tasks for all projects",
+  },
+  UpcomingTask: {
+    title: "Upcoming Tasks",
+    display: GetUpcomingTaskDisplay,
+    linkRoute: Routes.AllTasksPage(),
+    linkText: "All Tasks",
+    size: "col-span-6",
+    tooltipId: "tool-upcoming",
+    tooltipContent: "Three upcoming tasks for all projects",
+  },
+  ProjectSummary: {
+    title: project.name,
+    display: GetProjectSummaryDisplay,
+    linkRoute: Routes.EditProjectPage({ projectId: project.id }),
+    linkText: "Edit Project",
+    size: "col-span-6",
+    tooltipId: "tool-project",
+    tooltipContent: "Overall project information",
+  },
+  ProjectNotifications: {
+    title: "Notifications",
+    display: GetNotificationDisplay,
+    linkRoute: Routes.ProjectNotificationsPage({ projectId: project.id }),
+    linkText: "Project Notifications",
+    size: "col-span-6",
+    tooltipId: "tool-notification",
+    tooltipContent: "Three notifications for this project",
+  },
+  ProjectOverdueTask: {
+    title: "Overdue Tasks",
+    display: GetProjectOverdueTaskDisplay,
+    linkRoute: Routes.TasksPage({
+      projectId: project.id,
+    }),
+    linkText: "All Tasks",
+    size: "col-span-6",
+    tooltipId: "tool-overdue",
+    tooltipContent: "Three overdue tasks for this project",
+  },
+  ProjectUpcomingTask: {
+    title: "Upcoming Tasks",
+    display: GetProjectUpcomingTaskDisplay,
+    linkRoute: Routes.TasksPage({
+      projectId: project.id,
+    }),
+    linkText: "All Tasks",
+    size: "col-span-6",
+    tooltipId: "tool-upcoming",
+    tooltipContent: "Three upcoming tasks for this project",
+  },
+  ContributorNumber: {
+    title: "Contributors",
+    display: GetContributorDisplay,
+    linkRoute: Routes.ContributorsPage({ projectId: project.id }),
+    linkText: "View",
+    size: "col-span-2",
+    tooltipId: "tool-contributors",
+    tooltipContent: "Total number of contributors",
+  },
+  TeamNumber: {
+    title: "Teams",
+    display: GetTeamDisplay,
+    linkRoute: Routes.TeamsPage({ projectId: project.id }),
+    linkText: "View",
+    size: "col-span-2",
+    tooltipId: "tool-teams",
+    tooltipContent: "Total number of teams",
+  },
+  FormNumber: {
+    title: "Forms",
+    display: GetFormDisplay,
+    linkRoute: Routes.MetadataPage({ projectId: project.id }),
+    linkText: "View",
+    size: "col-span-2",
+    tooltipId: "tool-forms",
+    tooltipContent: "Percent of forms completed",
+  },
+  TaskTotal: {
+    title: "Tasks",
+    display: GetTotalTaskDisplay,
+    linkRoute: Routes.TasksPage({
+      projectId: project.id,
+    }),
+    linkText: "View",
+    size: "col-span-2",
+    tooltipId: "tool-tasks",
+    tooltipContent: "Percent of tasks completed",
+  },
+  ElementSummary: {
+    title: "Elements",
+    display: GetElementDisplay,
+    linkRoute: Routes.ElementsPage({ projectId: project.id }),
+    linkText: "View",
+    size: "col-span-2",
+    tooltipId: "tool-element",
+    tooltipContent: "Number of elements for this project",
+  },
+  LabelsSummary: {
+    title: "Roles",
+    display: GetLabelsDisplay,
+    linkRoute: Routes.CreditPage({ projectId: project.id }),
+    linkText: "View",
+    size: "col-span-2",
+    tooltipId: "tool-labels",
+    tooltipContent: "Percent of contributors or tasks labeled",
+  },
+}
